requirements.txt
asgiref
certifi
cffi
charset-normalizer
colorama
cryptography
defusedxml
Django
django-allauth
django-filter
django-rest-passwordreset
djangorestframework
idna
iniconfig
Markdown
oauthlib
packaging
pluggy
psycopg2-binary
pycparser
PyJWT
pytest
pytest-django
python-dotenv
python3-openid
pytz
PyYAML
requests
requests-oauthlib
sqlparse
tzdata
ujson
urllib3
gunicorn